pyproject.toml
[tool.poetry]
name = "np-team-task1-openapi"
version = "0.1.0"
description = ""
authors = ["mark47B <windachae47@gmail.com>"]
readme = "README.md"
packages = [{include = "np_team_task1_openapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
google-api-python-client = "^2.95.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
uvicorn = "^0.23.1"
pydantic = "^2.1.1"
oauth2client = "^4.1.3"
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.4.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
